requirements.txt
transformers==4.30.2
torch==2.0.1
neo4j==5.11.0
# deepspeech==0.9.3
# tacotron2==0.1.0
sentence-transformers==2.2.2
huggingface_hub==0.15.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
fastapi==0.100.0
uvicorn==0.23.1
python-multipart==0.0.6
pandas
scikit-learn
torch
seqeval